requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.28.0
email-validator==2.0.0

# Celery and task management
celery==5.3.4
redis==5.0.1
flower==2.0.1
apscheduler==3.10.4

# Web scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.8.5
async-timeout>=4.0.0,<5.0.0

# Data processing and analysis
pandas==1.5.3
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.3.0
spacy==3.6.1
transformers==4.30.2

# Email and notifications
jinja2==3.1.2
aiosmtplib==2.0.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Monitoring and logging
prometheus-client==0.17.1
sentry-sdk==1.32.0
loguru==0.7.2

# Utilities
python-dotenv==1.0.0
setuptools>=61.0.0 